+++ CHANGELOG.md
@@ -110,6 +110,7 @@
 * [X] 提示定义的本地化名字（对应的localisation.name为name或title）
 * [X] 修复cwt规则文件解析问题（有时无法解析optionComment中的`<>`）
 * [X] 即使cwt规则文件中没写也为definition补充某些definitionLocalisation，从已有的推断（如果有的话）（name title desc effect） 
+* [X] 实现LineMarkerProviderDescriptor改为实现RelatedItemLineMarkerProvider，从而实现`Go to related symbols`
 * [ ] scriptProperty的propertyName和definitionName不一致导致重命名scriptProperty时出现问题
 * [ ] 支持基于cwt配置文件进行脚本文件代码提示
 * [ ] 支持基于cwt配置文件进行脚本文件代码验证